Fix Apple Pay header button Liquid Glass corner radius (#5712)
## Summary
Previously, this corner radius was set to a constant 34. However, this
value was too high, and we didn't notice because under some
circumstances iOS only used that as a max and didn't actually set it
that high. It sometimes does use the full value, though, resulting in
broken appearance. This change it to instead use half the button's
height. This aligns with the `uniform` corner configuration API that the
Link button uses.

## Motivation
Fix broken UI

## Testing
Before:

![IMG_3773](https://github.com/user-attachments/assets/c8c260b1-07d2-4ed6-98b6-f73efb714adf)
After:

![IMG_3775](https://github.com/user-attachments/assets/9b9b446e-c303-4b11-8218-4069b54e2b20)
Theoretical appearance if we changed the size of the buttons (they're
currently fixed):

![IMG_3774](https://github.com/user-attachments/assets/849762a1-f347-44df-a41b-7bfa7053320a)

## Changelog
N/A