Send client_attribution_metadata on all confirmation calls + tests (#5732)
## Summary
<!-- Simple summary of what was changed. -->
Added client_attribution_metadata to confirm call for Apple Pay and
confirmation tokens (already taken care of the rest) and added network
stub testing with stub verification.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Previously, the expected behavior was that if payment method creation
and intent confirmation happen on separate requests in the same button
tap (like with deferred), include the metadata in payment method
creation rather than intent confirmation.

But now the data team wants client_attribution_metadata sent on both in
order to make it explicit that the intent-confirm and PM-create happened
on the same session.

https://stripe.slack.com/archives/C07PXHWQMP0/p1760553051428099?thread_ts=1760449818.955979&cid=C07PXHWQMP0
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Added stub testing with verification that the stubs contain the correct
metadata for new pms and saved pms for all intent and confirmation
flows. Couldn't figure out how to do it with Apple Pay or Link because
they have UI that needs interaction to confirm.

Apple Pay confirm request:
https://admin.corp.stripe.com/request-log/req_J5HL6cBFivDLLT
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A