Add placeholder for missing localized string (#5754)
## Summary

Adds a placeholder for the missing Chinese Traditional string in the
crypto onramp SDK. Also put it a request to have this string localized
https://jira.corp.stripe.com/browse/L10N-7764 which should override this
value once it's ready.

## Motivation

Release script has been bothering us that this is missing.

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->