CryptoOnramp SDK: Removes unused `Calendar` references (#5757)
## Summary
A while back, we simplified how dates of birth were being
encoded/decoded and stored. That was done here:
https://github.com/stripe/stripe-ios/pull/5351

Today, I noticed some lingering references to a `Calendar` being
specified, but not used. This removes that unused code.

## Motivation
Cleanup of unused code.

## Testing
Made sure the project built and the existing tests still passed.

## Changelog
N/A