Sort swiftinterface files when diffing (#5766)
## Summary
The API checker has false positives if the ordering of the API functions
in the `.swiftinterface` file changes, and Xcode seems to be compiling
these files in a non-deterministic order now that we're using folder
references. To work around this, try sorting them all before diffing.
This makes the files semantically meaningless, but it *does* catch
changes!

## Motivation
Fixing the API checker

## Testing
CI

## Changelog
N/A