CryptoOnramp SDK: Moves More KYC Refresh Logic to PaymentSheet (#5801)
## Summary
This PR moves more of the presentation and dismissal logic for KYC
refresh to `LinkController`. This allows us to remove `@_spi(STP)
public` exposure of the `VerifyKYCViewController`, and the other
cascading exposures due to this view controller being a subclass of the
previously unexposed `BottomSheetViewController` seen in #5787.

## Motivation
In response to comments on #5787, and to reduce the amount of code
exposed between frameworks to only what's absolutely necessary.

## Testing
1. Tested error throwing from the first (retreive) API call, when a user
doesn’t have previously attached KYC info:


https://github.com/user-attachments/assets/b2cb0f9d-fea4-4c94-b17c-eeb0be258201

2. Tested all other scenarios shown in the video below:
    1. Cancellation on the presented sheet.
    2. Confirmation without changes on the presented sheet.
    3. Changing address and confirming successfully.
4. Changing address and confirming unsuccessfully (using a NY address,
which is considered invalid by the confirmation API)


https://github.com/user-attachments/assets/1b5ba099-0f98-4237-9608-acdae5eab338

## Changelog
N/A