Create New iOS Model for Apple Issuing Push Provision P1 (#5779)
## Summary
<!-- Simple summary of what was changed. -->

This is part 1 of migrating the Issuing Push Provision library to a
standalone module.

Following the guidelines for setting up a new SDK module, I am creating
a new model for the Stripe iOS SDK.


## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
RUN_BAAS_CARDS-8638

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->