Update invalid placeholder in Link string (#5820)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes an issue with an invalid placeholder in the
`Enter the code sent to %@.` string.

The issue likely occurred because the Android repo sent a new string to
Lokalise without wrapping placeholders in `[%s]`, which caused the iOS
script to not replace `%s` with `%@`. This issue is being resolved
[here](https://github.com/stripe/stripe-android/pull/11956).

We're also working to fixing the string resources in Lokalise.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->


[RUN_LINK_MOBILE-52](https://jira.corp.stripe.com/browse/RUN_LINK_MOBILE-52)

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

Manually checked.

| Before | After |
|--------|--------|
| <img width="1206" height="2622" alt="image"
src="https://github.com/user-attachments/assets/dab6264c-3ebe-43e4-aea6-c4b2976865ef"
/> | <img width="1206" height="2622" alt="Simulator Screenshot - iPhone
17 Pro - 2025-11-11 at 14 29 10"
src="https://github.com/user-attachments/assets/bbda5c87-0f3b-41c4-9264-9ab020cd2b15"
/> |

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->