Attestation on confirmation (#5762)
## Summary
<!-- Simple summary of what was changed. -->
Add an AttestationChallenge wrapper around StripeAttest that prewarms
attestation on init and fetches an assertion with a timeout.

Wrap PassiveCaptchaChallenge and AttestationChallenge in
ConfirmationChallenge. Anywhere we previously passed
PassiveCaptchaChallenge, we now pass in ConfirmationChallenge.
ConfirmationChallenge creates the passive captcha and attestation
challenges asynchronously in parallel and fetches their tokens with
timeout in parallel and creates the radar options for the requests.

Added playground switch so that we can test with/without attestation,
similar to the passive captcha switch. Will be removed on release.

Moved PassiveCaptchaChallenge into StripePaymentSheet, @_spi(STP)
HCaptcha
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Attestation support for HCaptcha project
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Added tests
Example request:
https://admin.corp.stripe.com/request-log/req_YXMjaq42sec9gb
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A