Active captcha (#5819)
## Summary
<!-- Simple summary of what was changed. -->
Run Stripe js on a webview for active captcha.

For reference, this is the webview bridge for the challenge:
https://git.corp.stripe.com/stripe-internal/pay-server/blob/master/frontend/ocs-mobile/mobile-confirmation-challenge/src/components/webview.tsx
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
HCaptcha project
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Test using the "stripe_shop_test" merchant and the card `4000 0000 0000
1208`.


https://github.com/user-attachments/assets/e65d2041-cf63-4229-81c9-ce262ab6546b

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A