Replace test from `ConsumerSessionTests` with custom lint rule (#5857)
## Summary
<!-- Simple summary of what was changed. -->

This pull request removes the disabled
`testLookupSession_ignoresUnknownPaymentMethodType`, whose purpose was
to ensure that we handle unknown enum types in
`DisplayablePaymentDetails` gracefully, and replaces it with a custom
Swiftlint rule that enforces usage of `SafeEnumDecodable` instead of
`Decodable`. This will prevent any issues going forward.

Example run with forced failure:
https://app.bitrise.io/build/8dfe1bdd-2338-413e-8e8b-d71a4cc466c9

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

Safe enum decoding.

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

N/A

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->

N/A