Update test customer (#5866)
## Summary
<!-- Simple summary of what was changed. -->
Created a new customer and payment method for SEPA test. Rewrote save as
default test to create a new customer to attach pms to instead of
attaching to the SEPA customer.
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
The hardcoded customer hit the max amt of attached PMs allowed.
https://jira.corp.stripe.com/browse/RUN_MOBILESDK-4904
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Updated tests
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A