Add secure webview implementation of FC Lite for Instant Debits (#5869)
Related backend PRs:
[#1275019](https://git.corp.stripe.com/stripe-internal/pay-server/pull/1275019)
and
[#1280280](https://git.corp.stripe.com/stripe-internal/pay-server/pull/1280280)

## Summary

This updates FC Lite to use a secure webview instead of a `WKWebView`
for unverified merchants using Instant Debit. The rational behind this
is discussed in the Compass project / doc below. Note that


## Motivation


[Compass](https://home.corp.stripe.com/compass/projects/instant-debits-mobile-expansion)


[Doc](https://docs.google.com/document/d/1ouIrcJWZWKGf105JSoRP80UgcgQGJ5vRGH64K3H5wBo/edit?usp=sharing)

## Testing

FCLiteUITests cover these changes. Here is the new instant debits flow,
including a OAuth redirect:


https://github.com/user-attachments/assets/53ebc587-a38d-495a-8381-e8f4766971cb

Here is the FC Lite instant debits flow verified merchants would see
(using the `WKWebView`):


https://github.com/user-attachments/assets/f6aa3042-ddec-4546-aef0-1c802c974384

And the ACH flow still works with the WKWebView implementation:


https://github.com/user-attachments/assets/f3ea6026-f603-4476-91c2-c27416dfee30

## Changelog

N/a