CryptoOnramp SDK: More API Test Coverage (#5881)
## Summary
This PR adds API success / failure tests for the methods in
`STPAPIClient+CryptoOnramp.swift` that aren't already covered.

## Motivation
To complete the test coverage of these API methods, and improve overall
test coverage.

## Testing
Ran the full set of tests and verified that they successfully passed.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A