CryptoOnramp SDK: Stabilizes End-to-end UI Test (#5896)
## Summary
When transitioning between steps in the example app flow, network
operations occur which display a loading indicator that disables buttons
and blocks the UI. One of the steps was occasionally failing due to the
loading indicator still being on screen when the UI test attempted to
tap a button. See the Motivation section below for more info on the
specific failure.

This PR adds additional steps when transitioning between views to ensure
that the loading view has disappeared before attempting to tap buttons.

## Motivation

- To reduce sporadic test failures on CI.
- Specifically, to address the issue brought up and discussed here:
https://github.com/stripe/stripe-ios/pull/5879#discussion_r2604034639

## Testing
1. Ran the test target multiple times locally to ensure the modified
test was passing.
2. Temporarily introduced a deliberate delay of the removal of the
loading indicator at the failing test site, in `PaymentView.swift`.
```diff
private func fetchPaymentTokens() {
    isLoading.wrappedValue = true
    Task {
        do {
            let response = try await APIClient.shared.fetchPaymentTokens()
-           await MainActor.run {
+           DispatchQueue.main.asyncAfter(deadline: .now() + 3) {
                isLoading.wrappedValue = false
                paymentTokens = response.data
            }
        } catch {
            await MainActor.run {
                isLoading.wrappedValue = false
            }
        }
    }
}
```
3. Ran the test without the fix in place and made sure it failed as
expected.
4. Re-applied the fix.
5. Ran the test again and made sure it passed.

## Changelog
N/A