Add StripeIssuing to CI (#5899)
## Summary
<!-- Simple summary of what was changed. -->
Followed
[guide](https://confluence.corp.stripe.com/spaces/MOBILE/pages/310463183/iOS+SDK+Runbook+Configuring+module+tests)
to add test suite to CI.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Create a new SDK for StripeIssuing

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Ran `bundle install && bundle exec fastlane stripeissuing_tests`

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
Not a user affecting change.

---------

Co-authored-by: David Estes <davidestes@stripe.com>