ConfirmButton refactor 2/4 - Align button "status" naming scheme (#5894)
## Summary
Part 2/4 of a factor of the ConfirmButton. Previously the ConfirmButton
class supported both the Apple Pay button and confirm buttons, but now
it only supports confirm buttons and can be simplified, improve code
clarity and SDK binary size.

This PR aligns the name of the button "status" throughout the codebase.
Previously a mix of "state" and "status" was used, causing confusion and
readability issues. I chose "status" since the "State" type and "state"
properties are already used by UIControl, making it confusing and
ambiguous

## Motivation
improve readability and code clarity, prepare for further refactoring

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->