ConfirmButton refactor 3/4 - flatten appearance code (#5923)
## Summary
Part 3/4 of a factor of the ConfirmButton. Previously the ConfirmButton
class supported both the Apple Pay button and confirm buttons, but now
it only supports confirm buttons and can be simplified, improve code
clarity and SDK binary size.
Draft PR showing the whole change:
https://github.com/stripe/stripe-ios/pull/5678.

This PR flattens the appearance code to live directly in the BuyButton.
This sets up the BuyButton implementation to replace the ConfirmButton
wrapper and itself become the ConfirmButton

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->